Add alert box functionality and implement user update checks

+++ movieproject-api/src/AdminController.php
@@ -52,14 +52,20 @@ class AdminController
         $data = (array) json_decode(file_get_contents("php://input"), true);
         $errors = $this->getRegistrationValidationErrors(($data));
 
+        if ($this->gateway->isUserExists($data['email'])) {
+            http_response_code(409);
+            echo json_encode(["message" => "Account already exists. Please log in!"]);
+            return;
+        }
+
         if (!empty($errors)) {
             http_response_code(422);
             echo json_encode(["errors" => $errors]);
         } else {
-            $id = $this->gateway->register($data); 
+            $id = $this->gateway->register($data);
             http_response_code(201);
             echo json_encode([
-                "message" => "User created",
+                "message" => "Registration Complete!",
                 "id" => $id
             ]);
         }
